refactor: api 폴더의 server action들 action 폴더로 이동
