+++ actions/noticeServer.ts
@@ -1,7 +1,7 @@
 import { NoticePreviewList, Notice } from '@/types/notice';
 import { PostSearchQueryParams } from '@/types/post';
 
-import { deleteRequest, getRequest, patchRequest } from './serverIndex';
+import { deleteRequest, getRequest, patchRequest } from '../apis/serverIndex';
 
 const noticePath = '/notice';
