+++ actions/seminarActions.ts
@@ -3,11 +3,10 @@
 import { revalidateTag } from 'next/cache';
 import { redirect } from 'next/navigation';
 
-import { deleteSeminar } from '@/apis/seminarServer';
+import { deleteSeminar } from '@/actions/seminarServer';
 
 import { getPath } from '@/utils/page';
 import { seminar } from '@/utils/segmentNode';
-
 
 const seminarPath = getPath(seminar);
