+++ app/[locale]/community/notice/[id]/page.tsx
@@ -1,4 +1,4 @@
-import { getNoticePostDetail } from '@/apis/noticeServer';
+import { getNoticePostDetail } from '@/actions/noticeServer';
 
 import AdjPostNav from '@/components/common/AdjPostNav';
 import Attachments from '@/components/common/Attachments';
