+++ components/reservations/modals/useAddReservation.ts
@@ -1,8 +1,8 @@
 import { useReducer, useState, FormEventHandler } from 'react';
 
+import { postReservation } from '@/actions/reservation';
 import { NetworkError } from '@/apis';
 
-import { postReservation } from '@/apis/reservation';
 
 import { ReservationPostBody } from '@/types/reservation';
